deleted chain,portfolio and utils python files for the hosting purpose and added in the main python file
